Encapsulate observer code in tryCatch (#171)

+++ R/mod_log.R
@@ -39,44 +39,42 @@ mod_log_server <- function(id, r, d, m, language, i18n, debug, user_accesses){
     # Create log file if doesn't exist
     if (!file.exists(log_file)) file.create(log_file)
     
-    observeEvent(shiny.router::get_page(), {
-      req(shiny.router::get_page() == "log")
-      if (debug) cat(paste0("\n", now(), " - mod_log - observer shiny.router::get_page()"))
-      shinyjs::runjs(paste0("Shiny.setInputValue('", id, "-refresh_log', Math.random());"))
-    })
+    observeEvent(shiny.router::get_page(), try_catch("shiny.router::get_page()", {
+      if (shiny.router::get_page() == "log") shinyjs::runjs(paste0("Shiny.setInputValue('", id, "-refresh_log', Math.random());"))
+    }))
     
-    observeEvent(input$refresh_log, {
-      if (debug) cat(paste0("\n", now(), " - mod_log - observer input$refresh_log"))
+    observeEvent(input$refresh_log, try_catch("input$refresh_log", {
       
-      req("log_user_log" %in% user_accesses)
+      if ("log_user_log" %in% user_accesses){
       
-      log <-
-        readLines(log_file, warn = FALSE) %>% 
-        gsub("^(.*?\\- )(.*?)( \\- .*)$", "\\1<span style='color: #015bb5; font-weight: 600;'>\\2</span>\\3", .) %>%
-        gsub("Error", "<span style='color: #f44336; font-weight: 600;'></span>", .) %>%
-        # rev() %>% 
-        paste(collapse = "<br />")
+        log <-
+          readLines(log_file, warn = FALSE) %>% 
+          gsub("^(.*?\\- )(.*?)( \\- .*)$", "\\1<span style='color: #015bb5; font-weight: 600;'>\\2</span>\\3", .) %>%
+          gsub("Error", "<span style='color: #f44336; font-weight: 600;'></span>", .) %>%
+          # rev() %>% 
+          paste(collapse = "<br />")
+          
+        output$log <- renderUI(div(HTML(log), style = "font-weight: 500; font-family: monospace;"))
         
-      output$log <- renderUI(div(HTML(log), style = "font-weight: 500; font-family: monospace;"))
-      
-      # Scroll to the bottom of the page
-      shinyjs::delay(100,
-        shinyjs::runjs("
-          const element = document.getElementById('log-main');
-          element.scrollTop = element.scrollHeight;")
-      )
-    })
+        # Scroll to the bottom of the page
+        shinyjs::delay(100,
+          shinyjs::runjs("
+            const element = document.getElementById('log-main');
+            element.scrollTop = element.scrollHeight;")
+        )
+      }
+    }))
     
-    observeEvent(input$reset_log, {
-      if (debug) cat(paste0("\n", now(), " - mod_log - observer input$reset_log"))
+    observeEvent(input$reset_log, try_catch("input$reset_log", {
       
-      req("log_user_log" %in% user_accesses)
+      if ("log_user_log" %in% user_accesses){
       
-      file.remove(log_file)
-      file.create(log_file)
-      sink(log_file, append = TRUE)
-      
-      output$log <- renderUI("")
-    })
+        file.remove(log_file)
+        file.create(log_file)
+        sink(log_file, append = TRUE)
+        
+        output$log <- renderUI("")
+      }
+    }))
   })
 }
